ensure feature gating
