+++ src/lib.rs
@@ -27,7 +27,7 @@ pub fn get_app_version() -> &'static Version {
     })
 }
 
-#[cfg(feature = "mqtt")]
+#[cfg(all(not(target_arch = "wasm32"), feature = "mqtt"))]
 pub mod mqtt;
 pub mod plot;
 #[cfg(all(feature = "profiling", not(target_arch = "wasm32")))]
