+++ src/mqtt.rs
@@ -4,7 +4,6 @@ use egui::ScrollArea;
 use egui::Ui;
 use mqtt::data_receiver::MqttDataReceiver;
 use mqtt::mqtt_cfg_window::MqttConfigWindow;
-use mqtt::MqttPoint;
 
 fn show_broker_status(ui: &mut Ui, broker_status: Option<&Result<(), String>>) {
     if let Some(status) = broker_status {
