+++ src/plot/plot_graphics.rs
@@ -2,14 +2,9 @@ use egui::{Vec2, Vec2b};
 use egui_plot::{AxisHints, HPlacement, Legend, Plot, PlotBounds};
 use plot_util::{PlotData, Plots};
 
-use crate::plot::util;
+use crate::plot::{util, PlotMode};
 
 use super::{axis_config::AxisConfig, plot_settings::PlotSettings, ClickDelta, PlotType};
-
-enum PlotMode {
-    Logs,
-    MQTT,
-}
 
 /// Paints multiple plots based on the provided settings and configurations.
 ///
@@ -31,27 +26,21 @@ enum PlotMode {
 pub fn paint_plots(
     ui: &mut egui::Ui,
     reset_plot_bounds: bool,
-    plots: &mut Plots,
     plot_settings: &PlotSettings,
     legend_cfg: &Legend,
     axis_cfg: &AxisConfig,
     link_group: egui::Id,
     line_width: f32,
     click_delta: &mut ClickDelta,
-    mqtt_plots: &[mqtt::MqttPoints],
-    auto_scale: &mut bool,
+    mode: PlotMode<'_>,
 ) {
     #[cfg(all(feature = "profiling", not(target_arch = "wasm32")))]
     puffin::profile_function!();
-    let mode = if mqtt_plots.is_empty() {
-        PlotMode::Logs
-    } else {
-        PlotMode::MQTT
-    };
+
     let x_axes = vec![AxisHints::new_x().formatter(crate::util::format_time)];
 
     match mode {
-        PlotMode::Logs => {
+        PlotMode::Logs(plots) => {
             let plot_height = ui.available_height() / (plot_settings.total_plot_count() as f32);
 
             let percentage_plot = build_plot_ui(
@@ -110,7 +99,8 @@ pub fn paint_plots(
                 click_delta,
             );
         }
-        PlotMode::MQTT => {
+        #[cfg(all(not(target_arch = "wasm32"), feature = "mqtt"))]
+        PlotMode::MQTT(mqtt_plots, auto_scale) => {
             let mqtt_plot = build_plot_ui(
                 "mqtt",
                 ui.available_height(),
@@ -119,7 +109,7 @@ pub fn paint_plots(
                 x_axes,
                 link_group,
             );
-            fill_mqtt_plots(
+            crate::plot::plot_mqtt::fill_mqtt_plots(
                 ui,
                 reset_plot_bounds,
                 line_width,
@@ -196,131 +186,6 @@ fn fill_log_plots(
             fill_plot(plot_ui, plot, line_width, plot_settings);
         });
     }
-}
-
-/// Iterates through and fills/paints all plots with their respective data.
-///
-/// # Arguments
-///
-/// * `gui` - The egui UI to paint on.
-/// * `reset_plot_bounds` - whether plot bounds should be reset.
-/// * `line_width` - The width of plot lines.
-/// * `click_delta` - State relating to pointer clicks on plots
-fn fill_mqtt_plots(
-    gui: &mut egui::Ui,
-    reset_plot_bounds: bool,
-    line_width: f32,
-    click_delta: &mut ClickDelta,
-    mqtt_plot_area: Plot<'_>,
-    mqtt_plots: &[mqtt::MqttPoints],
-    auto_scale: &mut bool,
-) {
-    #[cfg(all(feature = "profiling", not(target_arch = "wasm32")))]
-    puffin::profile_function!();
-
-    let (scroll, modifiers) = util::get_cursor_scroll_input(gui);
-    let final_zoom_factor: Option<Vec2> = scroll.and_then(|s| util::set_zoom_factor(s, modifiers));
-
-    mqtt_plot_area.show(gui, |plot_ui| {
-        if plot_ui.response().hovered() {
-            if let Some(final_zoom_factor) = final_zoom_factor {
-                plot_ui.zoom_bounds_around_hovered(final_zoom_factor);
-            }
-        }
-        let resp = plot_ui.response();
-        if plot_ui.response().double_clicked() || reset_plot_bounds {
-            log::info!("Auto scaling re-enabled");
-            *auto_scale = true;
-            if let Some(max_bounds) = get_mqtt_auto_scaled_plot_bounds(mqtt_plots) {
-                plot_ui.set_plot_bounds(max_bounds);
-            }
-        } else if resp.clicked() {
-            *auto_scale = false;
-            log::info!("Auto scaling DISABLED!");
-            if plot_ui.ctx().input(|i| i.modifiers.shift) {
-                if let Some(pointer_coordinate) = plot_ui.pointer_coordinate() {
-                    click_delta.set_next_click(pointer_coordinate, PlotType::Hundreds);
-                }
-            } else {
-                click_delta.reset();
-            }
-        } else if resp.is_pointer_button_down_on() {
-            log::info!("Auto scaling DISABLED!");
-            *auto_scale = false;
-        } else if *auto_scale {
-            log::info!("Auto scaling enabled");
-            if let Some(max_bounds) = get_mqtt_auto_scaled_plot_bounds(mqtt_plots) {
-                plot_ui.set_plot_bounds(max_bounds);
-            }
-        }
-        click_delta.ui(plot_ui, PlotType::Hundreds);
-        let (x_lower, x_higher) = plot_util::extended_x_plot_bound(plot_ui.plot_bounds(), 0.1);
-        for mp in mqtt_plots {
-            if mp.data.len() < 2 {
-                // We don't plot less than two points. It's mostly because when the
-                // plotting starts, the auto-bounds causes a crash due to auto sizing
-                // a plot to 1 point and triggering an assert in egui_plot that the
-                // height and width of the bounds is greater than 0.0
-                continue;
-            }
-            plot_util::plot_raw_mqtt(
-                plot_ui,
-                &mp.topic,
-                &mp.data,
-                line_width,
-                (x_lower, x_higher),
-            );
-        }
-    });
-}
-
-fn get_mqtt_auto_scaled_plot_bounds(mqtt_plots: &[mqtt::MqttPoints]) -> Option<PlotBounds> {
-    let mut max_bounds: Option<PlotBounds> = None;
-    for mp in mqtt_plots {
-        let mp_first_point = mp
-            .data
-            .first()
-            .expect("Invalid empty MQTT plot data vector");
-        let tmp_bounds = if mp.data.len() < 2 {
-            PlotBounds::from_min_max(
-                [mp_first_point.x, mp_first_point.y],
-                [mp_first_point.x, mp_first_point.y],
-            )
-        } else {
-            let min_x = mp_first_point.x;
-            let max_x = mp
-                .data
-                .last()
-                .expect("Should be unreachable: Invalid empty MQTT plot data vector")
-                .x;
-            let mut min_y = mp_first_point.y;
-            let mut max_y = mp_first_point.y;
-            for p in &mp.data {
-                if p.y < min_y {
-                    min_y = p.y;
-                }
-                if p.y > max_y {
-                    max_y = p.y;
-                }
-            }
-            PlotBounds::from_min_max([min_x, min_y], [max_x, max_y])
-        };
-        if let Some(max_bounds) = &mut max_bounds {
-            max_bounds.merge(&tmp_bounds);
-        } else {
-            max_bounds = Some(tmp_bounds);
-        }
-    }
-    if let Some(mut max_bounds) = max_bounds {
-        // finally extend each bound by 10%
-        let margin_fraction = egui::Vec2::splat(0.1);
-        max_bounds.add_relative_margin_x(margin_fraction);
-        max_bounds.add_relative_margin_y(margin_fraction);
-        if max_bounds.is_valid() {
-            return Some(max_bounds);
-        }
-    }
-    None
 }
 
 /// Fills and paints a single plot with its data.
